+++ server/src/internal/song/song_test.go
@@ -6,16 +6,16 @@ import (
 	"github.com/labstack/echo/v4"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/veedubyou/chord-paper-be/server/src/errors/auth"
-	"github.com/veedubyou/chord-paper-be/server/src/errors/gateway"
-	. "github.com/veedubyou/chord-paper-be/server/src/lib/testing"
-	authtest "github.com/veedubyou/chord-paper-be/server/src/shared_tests/auth"
-	songerrors "github.com/veedubyou/chord-paper-be/server/src/song/errors"
-	songgateway "github.com/veedubyou/chord-paper-be/server/src/song/gateway"
-	songstorage "github.com/veedubyou/chord-paper-be/server/src/song/storage"
-	songusecase "github.com/veedubyou/chord-paper-be/server/src/song/usecase"
-	userstorage "github.com/veedubyou/chord-paper-be/server/src/user/storage"
-	userusecase "github.com/veedubyou/chord-paper-be/server/src/user/usecase"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/errors/auth"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/errors/gateway"
+	. "github.com/veedubyou/chord-paper-be/server/src/internal/lib/testing"
+	authtest "github.com/veedubyou/chord-paper-be/server/src/internal/shared_tests/auth"
+	songerrors "github.com/veedubyou/chord-paper-be/server/src/internal/song/errors"
+	songgateway "github.com/veedubyou/chord-paper-be/server/src/internal/song/gateway"
+	songstorage "github.com/veedubyou/chord-paper-be/server/src/internal/song/storage"
+	songusecase "github.com/veedubyou/chord-paper-be/server/src/internal/song/usecase"
+	userstorage "github.com/veedubyou/chord-paper-be/server/src/internal/user/storage"
+	userusecase "github.com/veedubyou/chord-paper-be/server/src/internal/user/usecase"
 	"net/http"
 	"net/http/httptest"
 	"time"
