+++ server/src/internal/track/entity/tracklist.go
@@ -3,7 +3,7 @@ package trackentity
 import (
 	"github.com/cockroachdb/errors"
 	"github.com/google/uuid"
-	"github.com/veedubyou/chord-paper-be/server/src/lib/jsonlib"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/lib/jsonlib"
 )
 
 const (
