+++ server/src/internal/track/track_suite_test.go
@@ -2,8 +2,8 @@ package track_test
 
 import (
 	"github.com/rabbitmq/amqp091-go"
-	dynamolib "github.com/veedubyou/chord-paper-be/server/src/lib/dynamo"
-	testlib "github.com/veedubyou/chord-paper-be/server/src/lib/testing"
+	dynamolib "github.com/veedubyou/chord-paper-be/server/src/internal/lib/dynamo"
+	testlib "github.com/veedubyou/chord-paper-be/server/src/internal/lib/testing"
 	"testing"
 
 	. "github.com/onsi/ginkgo/v2"
