+++ server/src/internal/track/track_test.go
@@ -6,21 +6,21 @@ import (
 	"github.com/labstack/echo/v4"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/veedubyou/chord-paper-be/server/src/lib/jsonlib"
-	"github.com/veedubyou/chord-paper-be/server/src/lib/rabbitmq"
-	. "github.com/veedubyou/chord-paper-be/server/src/lib/testing"
-	authtest "github.com/veedubyou/chord-paper-be/server/src/shared_tests/auth"
-	songerrors "github.com/veedubyou/chord-paper-be/server/src/song/errors"
-	songgateway "github.com/veedubyou/chord-paper-be/server/src/song/gateway"
-	songstorage "github.com/veedubyou/chord-paper-be/server/src/song/storage"
-	songusecase "github.com/veedubyou/chord-paper-be/server/src/song/usecase"
-	trackentity "github.com/veedubyou/chord-paper-be/server/src/track/entity"
-	trackerrors "github.com/veedubyou/chord-paper-be/server/src/track/errors"
-	trackgateway "github.com/veedubyou/chord-paper-be/server/src/track/gateway"
-	trackstorage "github.com/veedubyou/chord-paper-be/server/src/track/storage"
-	trackusecase "github.com/veedubyou/chord-paper-be/server/src/track/usecase"
-	userstorage "github.com/veedubyou/chord-paper-be/server/src/user/storage"
-	userusecase "github.com/veedubyou/chord-paper-be/server/src/user/usecase"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/lib/jsonlib"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/lib/rabbitmq"
+	. "github.com/veedubyou/chord-paper-be/server/src/internal/lib/testing"
+	authtest "github.com/veedubyou/chord-paper-be/server/src/internal/shared_tests/auth"
+	songerrors "github.com/veedubyou/chord-paper-be/server/src/internal/song/errors"
+	songgateway "github.com/veedubyou/chord-paper-be/server/src/internal/song/gateway"
+	songstorage "github.com/veedubyou/chord-paper-be/server/src/internal/song/storage"
+	songusecase "github.com/veedubyou/chord-paper-be/server/src/internal/song/usecase"
+	trackentity "github.com/veedubyou/chord-paper-be/server/src/internal/track/entity"
+	trackerrors "github.com/veedubyou/chord-paper-be/server/src/internal/track/errors"
+	trackgateway "github.com/veedubyou/chord-paper-be/server/src/internal/track/gateway"
+	trackstorage "github.com/veedubyou/chord-paper-be/server/src/internal/track/storage"
+	trackusecase "github.com/veedubyou/chord-paper-be/server/src/internal/track/usecase"
+	userstorage "github.com/veedubyou/chord-paper-be/server/src/internal/user/storage"
+	userusecase "github.com/veedubyou/chord-paper-be/server/src/internal/user/usecase"
 	"net/http"
 	"net/http/httptest"
 )
