+++ server/src/internal/user/storage/db.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"github.com/cockroachdb/errors/markers"
 	"github.com/guregu/dynamo"
-	dynamolib "github.com/veedubyou/chord-paper-be/server/src/lib/dynamo"
-	"github.com/veedubyou/chord-paper-be/server/src/lib/errors/mark"
-	userentity "github.com/veedubyou/chord-paper-be/server/src/user/entity"
+	dynamolib "github.com/veedubyou/chord-paper-be/server/src/internal/lib/dynamo"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/lib/errors/mark"
+	userentity "github.com/veedubyou/chord-paper-be/server/src/internal/user/entity"
 )
 
 const (
