+++ server/src/internal/user/usecase/usecase.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"github.com/cockroachdb/errors/markers"
 	"github.com/pkg/errors"
-	"github.com/veedubyou/chord-paper-be/server/src/errors/api"
-	"github.com/veedubyou/chord-paper-be/server/src/errors/auth"
-	userentity "github.com/veedubyou/chord-paper-be/server/src/user/entity"
-	"github.com/veedubyou/chord-paper-be/server/src/user/google_id"
-	"github.com/veedubyou/chord-paper-be/server/src/user/storage"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/errors/api"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/errors/auth"
+	userentity "github.com/veedubyou/chord-paper-be/server/src/internal/user/entity"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/user/google_id"
+	"github.com/veedubyou/chord-paper-be/server/src/internal/user/storage"
 	"strings"
 	"sync"
 )
